@@ -39,6 +39,7 @@ gmock_output_test.py
 
 """
 
+from io import open    # pylint: disable=redefined-builtin, g-importing-member
 import os
 import re
 import sys
@@ -152,10 +153,11 @@ def GetNormalizedCommandOutputAndLeakyTests(cmd):
 
 
 class GMockOutputTest(gmock_test_utils.TestCase):
+
   def testOutput(self):
     (output, leaky_tests) = GetNormalizedCommandOutputAndLeakyTests(COMMAND)
     golden_file = open(GOLDEN_PATH, 'rb')
-    golden = golden_file.read()
+    golden = golden_file.read().decode('utf-8')
     golden_file.close()
 
     # The normalized output should match the golden file.
